+++ src/components/common/Header/Header.jsx
@@ -7,12 +7,8 @@ import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
 import Mic from '@material-ui/icons/Mic';
 import Navigation from '@material-ui/icons/Navigation';
-import Home from '@material-ui/icons/Home';
-import SupervisorAccount from '@material-ui/icons/SupervisorAccount';
-import Hotel from '@material-ui/icons/Hotel';
-import Brightness3 from '@material-ui/icons/Brightness3';
 import  { useHistory } from 'react-router-dom';
-import ItemMenu from 'components/shared/menu/ItemMenu';
+import HeaderMenu from 'components/common/Header/HeaderMenu';
 import useTheme from 'features/contexts/Theme/UseTheme';
 
 const useStyles = makeStyles((theme) => ({
@@ -30,10 +26,10 @@ export default function Header() {
   const [anchorEl, setAnchorEl] = React.useState(null);
   const currentUrl = history.location.pathname;
 
-  const menuItems = [{title: 'דף הבית', param: '/', icon: <Home fontSize="small" />, current: currentUrl === '/', onClickMethod: (e) => redirect(e)}, 
-                     {title: 'עמוד ניהול', param: '/managePage', icon: <SupervisorAccount fontSize="small" />, current: currentUrl === '/managePage', onClickMethod: (e) => redirect(e)},
-                     {title: 'עץ', param: '/treeDemo', icon: <Hotel fontSize="small" />, current: currentUrl === '/treeDemo', onClickMethod: (e) => redirect(e)},
-                     {title: 'החלף רקע', param: '', icon: <Brightness3 fontSize="small" />, current: false, onClickMethod: (e) => changeTheme(e)}]
+  // const menuItems = [{title: 'דף הבית', param: '/', icon: <Home fontSize="small" />, current: currentUrl === '/', onClickMethod: (e) => redirect(e)}, 
+  //                    {title: 'עמוד ניהול', param: '/managePage', icon: <SupervisorAccount fontSize="small" />, current: currentUrl === '/managePage', onClickMethod: (e) => redirect(e)},
+  //                    {title: 'עץ', param: '/treeDemo', icon: <Hotel fontSize="small" />, current: currentUrl === '/treeDemo', onClickMethod: (e) => redirect(e)},
+  //                    {title: 'החלף רקע', param: '', icon: <Brightness3 fontSize="small" />, current: false, onClickMethod: (e) => changeTheme(e)}]
 
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
@@ -62,8 +58,8 @@ export default function Header() {
           <IconButton className={classes.Spacing} edge="end" color="inherit" aria-label="menu" onClick={handleClick}>
             <MenuIcon />
           </IconButton>
-          <ItemMenu items={menuItems} handleClose={handleClose} anchorEl={anchorEl} onClickMethod={redirect}/>
-
+          {/* <ItemMenu items={menuItems} handleClose={handleClose} anchorEl={anchorEl} onClickMethod={redirect}/> */}
+              <HeaderMenu handleClose={handleClose} anchorEl={anchorEl} redirect={redirect} changeTheme={changeTheme} currentUrl={currentUrl}/>
           <Typography variant="h6" className='header-title' >
             עוף החול נייטרו
           </Typography>
